+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project tries its best to use [Semantic Versioning](http://semver.org/)
 - ScrollReveal returns a non-operational instance when instantiated in unsupported browsers.
 - Library version is now a read-only instance member.
 - `options.viewFactor` clamps values outside of `0.0` to `1.0`.
+- ScrollReveal constructor now returns a singleton.
 
 ### Fixed
 - The `requestAnimationFrame` polyfill now reliably throttles callback invocations.